+++ SideWarsShared/Networking/ServerPlayerMovementPacket.cs
@@ -3,7 +3,7 @@
     public class ServerPlayerMovementPacket
     {
         public int Id { get; set; }
-        public sbyte Horizontal { get; set; }
+        public float Horizontal { get; set; }
         //public bool Jump { get; set; }
     }
 }
